requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.11.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coverage==7.1.0
crispy-bootstrap5==0.7
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-admin-honeypot-updated-2021==1.2.0
django-bootstrap-v5==1.0.11
django-countries==7.5.1
django-crispy-forms==2.0
django-ipware==5.0.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-resized==1.0.2
ecdsa==0.18.0
exceptiongroup==1.1.1
firebase-admin==6.1.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
msgpack==1.0.5
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.8
packaging==23.1
phonenumbers==8.13.11
Pillow==9.4.0
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.22.1
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.1.0
social-auth-core==4.4.0
soupsieve==2.4
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.9.2
